% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_escalation.R
\name{dose_escalation}
\alias{dose_escalation}
\title{Decision about dose escalation based on the toxicities (Bayesian Decision-Theoretic Design)}
\usage{
dose_escalation(d, a.pTox, b.pTox, target, alpha, eta)
}
\arguments{
\item{d}{the current dose combination}

\item{a.pTox}{parameter alpha for current prior distribution}

\item{b.pTox}{parameter beta for current prior distribution}

\item{target}{the target toxicity level}

\item{alpha}{penalty term if the current DLT probability is below the target DLT probability}

\item{eta}{penalty term if the current DLT probability is above the target DLT probability}
}
\value{
\code{dose_escalation()} returns the indexes of next dose allocation
}
\description{
Find the intervals of biological efficacious doses (BEDs) based on prior information, current MTD, and current data.
}
\details{
\code{dose_escalation()} calculate the values of utility function, the dose combination with greatest utility is the next dose allocation
         In this function, if the current dose combination is (r,s), next dose allocation is limited to dose combination (i,j) such that i <= r and j<= s, as well as the combinations (r+1,s) and (r,s+1).
}
\seealso{
Fan, S. K., Lee, B. L. and Lu, Y.: A Curve-free Bayesian Decision-theoretic Design for Phase Ia/Ib Trials Considering both Safety and Efficacy Outcomes. Statistics in Biosciences, 12(2), 146-166 (2020). \url{https://doi.org/10.1007/s12561-020-09272-5}
}
