% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbeds_CFBD.R
\name{findbeds_CFBD}
\alias{findbeds_CFBD}
\title{Find intervals of biological efficacious doses (BEDs)}
\usage{
findbeds_CFBD(mtd, n.eff, n.assign, a.pEff, b.pEff, E.min, gain.A, gain.AC, phi, lo)
}
\arguments{
\item{mtd}{the current maximum tolerated dose (MTD)}

\item{n.eff}{a list of no. of efficacies at all dose levels}

\item{n.assign}{a list of no. of patients at all dose levels}

\item{a.pEff}{parameter alpha for current prior distribution}

\item{b.pEff}{parameter beta for current prior distribution}

\item{E.min}{the minimum acceptable efficacy rate}

\item{gain.A}{the weight of reward that a BED is selected correctly; 1 is suggested}

\item{gain.AC}{the weight of reward that a non-BED is not selected correctly; 1 is suggested}

\item{phi}{the weight for how much more efficacious P_{E,i} is than E.min; 1 is suggested}

\item{lo}{the weight for how much less efficacious P_{E,i} is than E.min; 1 is suggested}
}
\value{
\code{findbeds()} returns (1) iL.B: the lower boundary of the interval of BEDs and; (2) iU.B: the upper boundary of the interval of BEDs and; (3) p.best: the posterior probability that interval B is acceptable
}
\description{
Find the intervals of biological efficacious doses (BEDs) based on prior information, current MTD, and current data.
}
\details{
\code{findbeds()} implement the following procedure:
         (1) find the interval with maximum utility A among all the possible intervals
         (2) find the subinterval B that maximizes the posterior probability
         In this function, only doses at or below MTD are investigated and we use real data rather than working data to address the BED interval.
}
\seealso{
Fan, S., Lee, B. L., & Lu, Y. (2020). A curve-free bayesian decision-theoretic design for phase Ia/Ib trials considering both safety and efficacy outcomes. \emph{Statistics in Biosciences}, 12(2), 146–166. \url{https://doi.org/10.1007/s12561-020-09272-5}
}
