% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmtd.R
\name{findmtd}
\alias{findmtd}
\title{Find the maximum tolerated dose (MTD)}
\usage{
findmtd(target, a.pTox, b.pTox, alpha, eta)
}
\arguments{
\item{target}{the target toxicity level}

\item{a.pTox}{parameter alpha for current prior distribution}

\item{b.pTox}{parameter beta for current prior distribution}

\item{alpha}{the weight of penalty of toxicity rate below target level; 1 is suggested}

\item{eta}{the weight of penalty of toxicity rate above target level; 1 is suggested}
}
\value{
\code{findmtd()} returns: (1) gain and (2) mtd
}
\description{
Find the maximum tolerated dose (MTD) based on prior information and current data. Use working data instead of real data for small samples
}
\seealso{
Fan, S., Lee, B. L., & Lu, Y. (2020). A curve-free bayesian decision-theoretic design for phase Ia/Ib trials considering both safety and efficacy outcomes. \emph{Statistics in Biosciences}, 12(2), 146–166. \url{https://doi.org/10.1007/s12561-020-09272-5}
         Fan SK, Lu Y, Wang YG. A simple Bayesian decision-theoretic design for dose-finding trials. \emph{Stat Med}, 2012 Dec 10;31(28):3719-30. doi: 10.1002/sim.5438. Epub 2012 Jul 5. PMID: 22763943. \url{https://doi.org/10.1002/sim.5438}
}
