% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oc_2agents_bayesian.R
\name{get_oc_2agents_bayesian}
\alias{get_oc_2agents_bayesian}
\title{Generate operating characteristics for the Bayesian Decision-Theoretic Design}
\usage{
get_oc_2agents_bayesian(pTox, target, T.max, n.min, n.max, 
n.sim, seed, var.ratio = 4, alpha = 1.2, eta = 1, r1 = 0.5, r2 = 0.95, type = 1)
}
\arguments{
\item{pTox}{a list of true toxicity probabilities at all dose levels}

\item{target}{the target toxicity level}

\item{T.max}{the absolute overly toxic level, usually .05 above target}

\item{n.min}{the minimal sample size}

\item{n.max}{the maximal sample size}

\item{n.sim}{the total number of trials to be simulated}

\item{seed}{the random seed for simulation}

\item{var.ratio}{an equivalent no. of patients contained in the prior information, chosen such that prior.n.mtd * target is larger than but close to 1}

\item{alpha}{the weight of penalty of toxicity rate below target level; 1.2 is suggested}

\item{eta}{the weight of penalty of toxicity rate above target level; 1 is suggested}

\item{r1}{the error rate for concluding inadmissible; .10 is suggested}

\item{r2}{the error rate for concluding admissible; .10 is suggested}

\item{type}{the type of simulation, 1. Simulation study; 2. Sensitivity analysis based on random error (underestimate); 3. Sensitivity analysis based on random error (overestimate); 4. Sensitivity analysis based on fixed error}
}
\value{
\code{get_oc()} returns the operating characteristics of the CFBD design as a list,
        including:
        (1) \code{$n}: the average sample size over all simulated trials
        (2) \code{$selection_prop}: the selection percentages of each toxicity level
        (3) \code{$percentFound}: the percentage of trials recommending BEDs
        (5) \code{$percentTox}: the percentage of in-trial toxicities
        (6) \code{$percentMTD}: the selection percentages of each dose combination
        (7) \code{$percentPatients}: patient allocation for all doses under CFBD
}
\description{
Obtain the operating characteristics of the Bayesian Decision-Theoretic Design
}
\seealso{
Fan, S., Lee, B. L., & Lu, Y. (2020). A curve-free bayesian decision-theoretic design for phase Ia/Ib trials considering both safety and efficacy outcomes. \emph{Statistics in Biosciences}, 12(2), 146–166. \url{https://doi.org/10.1007/s12561-020-09272-5}
}
