% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oc_2DLT.R
\name{get_oc_2DLT}
\alias{get_oc_2DLT}
\title{Generate operating characteristics of The Rule-Based Design for Agents with Non-Overlapping Dose-Limiting Toxicities}
\usage{
get_oc_2DLT(p.true, p.conditional1, p.conditional2, 
A, B, C, a.E, a.D, b.E, b.D, c.E, ntrial = 50000, seed = NULL)
}
\arguments{
\item{p.true}{a list of true DLT rates at all dose combination levels}

\item{p.conditional1}{The conditional probability of observing only DLT1, given that at least one DLT occurs and that DLT3 is absent}

\item{p.conditional2}{The conditional probability of observing only DLT2, given that at least one DLT occurs and that DLT3 is absent}

\item{A}{The sample size of first cohort of A+B+C design}

\item{B}{The sample size of first cohort of A+B+C design}

\item{C}{The sample size of first cohort of A+B+C design}

\item{a.E}{is the maximum number of DLTs needed out of A patients to to conclude the current dose is accepted;}

\item{a.D}{is the minimum number of DLTs needed out of A patients to to conclude the current dose is rejected;}

\item{b.E}{is the maximum number of DLTs needed out of B patients to to conclude the current dose is accepted;}

\item{b.D}{is the minimum number of DLTs needed out of B patients to to conclude the current dose is rejected;}

\item{c.E}{is the minimum number of DLTs needed out of C patients to to conclude the current dose is rejected, otherwise, the current dose will be accepted;}

\item{ntrial}{the total number of trials to be simulated}

\item{seed}{the random seed for simulation}
}
\value{
\code{get_oc_2DLT()} returns the operating characteristics of the proposted design as a list,
        including:
        (1) \code{$selection_prop}: the selection percentages of each toxicity level
        (2) \code{$recommended_prop}: the selection percentages of each dose combination
        (3) \code{$nDLT_list}: a list of the number of DLT outcomes of each trial
        (4) \code{$nPatient_list}: a list of sample size of each trial
}
\description{
Obtain the operating characteristics of The Rule-Based Design for Agents with Non-Overlapping Dose-Limiting Toxicities by simulating trials
}
\seealso{
Lee BL, Fan SK. A two-dimensional search algorithm for dose-finding trials of two agents. J Biopharm Stat. 2012;22(4):802-18. doi: 10.1080/10543406.2012.676587. PMID: 22651116.
}
