% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t.table.R
\name{get_t.table}
\alias{get_t.table}
\title{Return a table containing the number of DLT1, DLT2, and DLT3 based on the conditional probability specified by users}
\usage{
get_t.table(d, n_DLTs, p.conditional1, p.conditional2)
}
\arguments{
\item{d}{the current dose combination}

\item{n_DLTs}{the number of DLT outcomes}

\item{p.conditional1}{The conditional probability of observing only DLT1, given that at least one DLT occurs and that DLT3 is absent}

\item{p.conditional2}{The conditional probability of observing only DLT2, given that at least one DLT occurs and that DLT3 is absent}
}
\value{
\code{get_decision()} returns the result as a list,
        including:
        (1) \code{$t.table}: A matrix containing the number of DLT1, DLT2, and DLT3, respectively.
}
\description{
Return a table containing the number of DLT1, DLT2, and DLT3 based on the conditional probability specified by users
}
\seealso{
Lee BL, Fan SK. A two-dimensional search algorithm for dose-finding trials of two agents. J Biopharm Stat. 2012;22(4):802-18. doi: 10.1080/10543406.2012.676587. PMID: 22651116.
}
