% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low.R
\name{low}
\alias{low}
\title{Determine whether (i,j) is the largest acceptable one}
\usage{
low(p_hat, n1.dose, n2.dose, mtdp, i, j)
}
\arguments{
\item{p_hat}{a matrix containing the isotonic estimation of DLT probability of each dose level}

\item{n1.dose}{the number of dose level of agent 1}

\item{n2.dose}{the number of dose level of agent 2}

\item{mtdp}{the toxicity rate of MTD}

\item{i}{the current dose level of agent 1}

\item{j}{the current dose level of agent 2}
}
\value{
\code{low()} returns: 1 = no, 0 = yes
}
\description{
Determine whether (i,j) is the largest acceptable one
}
\seealso{
Fan SK, Venook AP, Lu Y. Design issues in dose-finding Phase I trials for combinations of two agents. J Biopharm Stat. 2009;19(3):509-23. doi: 10.1080/10543400902802433. PMID: 19384692.
}
