% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.dim.search.R
\name{one.dim.search}
\alias{one.dim.search}
\title{A recursion function that perform one dimensional test of the Rule-Based Design for Agents with Same Dose-Limiting Toxicities at dose combination (i,j)}
\usage{
one.dim.search(p_hat, n1.dose, n2.dose, n, p, t, r, i, j)
}
\arguments{
\item{p_hat}{a matrix containing the isotonic estimation of DLT probability of each dose level}

\item{n1.dose}{the number of dose level of agent 1}

\item{n2.dose}{the number of dose level of agent 2}

\item{n}{a matrix containing the total number of patients of each dose level}

\item{p}{a matrix containing the true DLT probability of each dose level}

\item{t}{a matrix containing the number of DLT outcomes of each dose level}

\item{r}{a matrix containing the rejection decision of each dose level}

\item{i}{the current dose level of agent 1}

\item{j}{the current dose level of agent 2}
}
\value{
\code{one.dim.search()} returns:
        (1) \code{$t}: a matrix containing the number of DLT outcomes of each dose level
        (2) \code{$n}: a matrix containing the total number of patients of each dose level
        (3) \code{$r}: a matrix containing the rejection decision of each dose level
}
\description{
A recursion function that perform one dimensional test of the Rule-Based Design for Agents with Same Dose-Limiting Toxicities at dose combination (i,j), update the two-dimension isotonic estimation, and return the results
}
\seealso{
Fan SK, Venook AP, Lu Y. Design issues in dose-finding Phase I trials for combinations of two agents. J Biopharm Stat. 2009;19(3):509-23. doi: 10.1080/10543400902802433. PMID: 19384692.
}
