% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Perform 2+1+3 test for the Rule-Based Design for Agents with Same Dose-Limiting Toxicities at dose combination (k,l), and return the results}
\usage{
test(n1.dose, n2.dose, n, p, t, r, k, l)
}
\arguments{
\item{n1.dose}{the number of dose level of agent 1}

\item{n2.dose}{the number of dose level of agent 2}

\item{n}{a matrix containing the total number of patients of each dose level}

\item{p}{a matrix containing the true DLT probability of each dose level}

\item{t}{a matrix containing the number of DLT outcomes of each dose level}

\item{r}{a matrix containing the rejection decision of each dose level}

\item{k}{the current dose level of agent 1}

\item{l}{the current dose level of agent 2}
}
\value{
\code{test()} returns:
        (1) \code{$t}: a matrix containing the number of DLT outcomes of each dose level
        (2) \code{$n}: a matrix containing the total number of patients of each dose level
        (3) \code{$r}: a matrix containing the rejection decision of each dose level
}
\description{
Perform 2+1+3 test for the Rule-Based Design for Agents with Same Dose-Limiting Toxicities at dose combination (k,l), and return the results
}
\seealso{
Fan SK, Venook AP, Lu Y. Design issues in dose-finding Phase I trials for combinations of two agents. J Biopharm Stat. 2009;19(3):509-23. doi: 10.1080/10543400902802433. PMID: 19384692.
}
