% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twodim.iso.R
\name{twodim.iso}
\alias{twodim.iso}
\title{Update isotonic estimation of each dose combination}
\usage{
twodim.iso(n1.dose, n2.dose, p_hat, t, n)
}
\arguments{
\item{n1.dose}{the number of dose level of agent 1}

\item{n2.dose}{the number of dose level of agent 2}

\item{p_hat}{a matrix containing the isotonic estimation of DLT probability of each dose level}

\item{t}{a matrix containing the number of DLT outcomes of each dose level}

\item{n}{a matrix containing the total number of patients of each dose level}
}
\value{
\code{two.dim.search()} returns: p_hat a matrix containing the isotonic estimation of DLT probability of each dose level
}
\description{
Update isotonic estimation of each dose combination
}
\seealso{
Fan SK, Venook AP, Lu Y. Design issues in dose-finding Phase I trials for combinations of two agents. J Biopharm Stat. 2009;19(3):509-23. doi: 10.1080/10543400902802433. PMID: 19384692.
}
