% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_data_2dim.R
\name{work_data_2dim}
\alias{work_data_2dim}
\title{Calculate the 2 dimensional working data for small sample to find MTD efficiently.}
\usage{
workdata(n.tox, n.assign)
}
\arguments{
\item{n.tox}{a list of no. of toxicities at all dose levels}

\item{n.assign}{a list of no. of patients at all dose levels}
}
\value{
\code{workdata()} returns (1) tox: the working number of toxicities and; (2) n: the working number of patients
}
\description{
Calculate the 2 dimensional working data for small sample to find MTD efficiently.
}
\examples{
work <- workdata(c(1,1),c(2,3))
}
\seealso{
Fan SK, Venook AP, Lu Y. Design issues in dose-finding Phase I trials for combinations of two agents. J Biopharm Stat. 2009;19(3):509-23. doi: 10.1080/10543400902802433. PMID: 19384692.
}
